Remove deprecated files and update theme structure
change to dawn theme

This commit removes several outdated files and sections from the theme, including critical CSS, various Liquid sections, and snippets that are no longer in use. Additionally, it updates the JSON templates for improved organization and compliance with Shopify's latest standards. These changes aim to streamline the theme and enhance performance for a better user experience.